requirements.txt
Django==1.3.1
PIL==1.1.7
argparse==1.2.1
distribute==0.6.24
django-debug-toolbar==0.9.4
django-profiles==0.2
django-registration==0.8
python-memcached==1.48
sorl-thumbnail==11.12
wsgiref==0.1.2
